Split file settings save and backup mixins.
